Split the project into independent modules.
Splitting the project into independent modules allows for better and more granular use of access control, and provides for a better overall project structure.

In order to cope with this change, two delegates were added for the input and output modules, so that the input module can convey information to the user (only used to alert about CapsLock state changes for now), and the output module can check for certain key states. Previously the output code had a direct dependency on the input code, so if I used the same design to make the input code depend on the output code, that would lead to a circular dependency which is just bad design.