Added handling code to prevent event taps from being disabled when the screen-reader is busy.
When performing tasks that block the thread on which input events are processed for long periods of time, the Quartz Event Services disables any existing event taps to prevent the computer from becoming unresponsive to input. This change adds some code to handle the special events that are sent to tap handlers and re-enables the event tap after resuming execution when it gets disabled due to a timeout.